backend/requirements.txt
fastapi
uvicorn
python-multipart
pydicom
pillow
requests
openai
